Update tailwind css style, inport the icon elements from react icon.
