--- src/App.css
@@ -1,65 +1,0 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-.search-icon {
-  font-size: 20px;
-}
-
-.search-container {
-  width: 100%;
-  padding: 5px;
-  margin-top: 80px;
-}
-
-.search-input {
-  width: 100%;
-  max-width: 500px;
-  height: 45px;
-  background-color: white;
-  padding: 5px;
-  font-weight: bold;
-}
-
-.ticket-container{
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin-top: 120px;
-  margin-left: auto;
-  margin-right: auto;
-  gap: 50px;
-  max-width: 500px;
-  
-}
-
-.ticket-block,.existing-ticket-block{
-  display: flex;
-  align-items: center;
-  background-color: white;
-}
-
-.ticket-block:hover,
-.existing-ticket-block:hover {
-  background-color: brown;
-}
-
-
-.ticket-title{
-  color: black;
-}
-
-.ticket-description{
-  color: black;
-}
-
-.ticket-icon{
-  font-size: 25px;
-}
-
-
-
