+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import './App.css'
-import SearchBar from './components/Searchbar'
-import WorkOrder from './components/Ticket'
+import SearchBar from './components/Searchbar/Searchbar'
+import WorkOrder from './components/Ticket/Ticket'
 
 function App() {
     return(
