--- src/components/Searchbar.tsx
@@ -1,8 +1,0 @@
-export default function SearchBar() {
-    return(
-        <div className="search-container">
-            <span className="search-icon">🔍</span>
-            <input className="search-input" type="search" placeholder="Seach for content..."/>
-        </div>
-    );
-}
