+++ src/index.css
@@ -1,15 +1,15 @@
-:root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
+@import "tailwindcss";
 
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: black;
+@theme {
+  --color-lightblue: #90CAF9;
+  --color-indigo: #5C6BC0;
+  --color-blueGray: #B0BEC5;
+  --color-amber: #FFCA28;
+  --color-lightGreen: #9CCC65
+}
 
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+body {
+  margin: 0;
+  padding: 0;
 }
